+++ src/lib/Components/Consignments/Screens/Overview.tsx
@@ -1,5 +1,5 @@
+import { Schema, screenTrack, Track, track as _track } from "lib/utils/track"
 import React from "react"
-
 import {
   AsyncStorage,
   Dimensions,
@@ -12,13 +12,15 @@ import {
   ViewProperties,
 } from "react-native"
 
-import ConsignmentBG from "../Components/ConsignmentBG"
-import { LargeHeadline, Subtitle } from "../Typography"
-
 import { ConsignmentMetadata, ConsignmentSetup, SearchResult } from "../"
+import SwitchBoard from "../../../NativeModules/SwitchBoard"
 import TODO from "../Components/ArtworkConsignmentTodo"
-
+import ConsignmentBG from "../Components/ConsignmentBG"
 import { Button, Row } from "../Components/FormElements"
+import createSubmission from "../Submission/create"
+import updateSubmission from "../Submission/update"
+import { uploadImageAndPassToGemini } from "../Submission/uploadPhotoToGemini"
+import { LargeHeadline, Subtitle } from "../Typography"
 import Artist from "./Artist"
 import FinalSubmissionQuestions from "./FinalSubmissionQuestions"
 import Location from "./Location"
@@ -27,13 +29,7 @@ import Provenance from "./Provenance"
 import SelectFromPhotoLibrary from "./SelectFromPhotoLibrary"
 import Welcome from "./Welcome"
 
-import createSubmission from "../Submission/create"
-import updateSubmission from "../Submission/update"
-import { uploadImageAndPassToGemini } from "../Submission/uploadPhotoToGemini"
-
 const consignmentsStateKey = "ConsignmentsStoredState"
-
-import SwitchBoard from "../../../NativeModules/SwitchBoard"
 
 interface Props extends ViewProperties {
   navigator: NavigatorIOS
@@ -45,6 +41,12 @@ interface State extends ConsignmentSetup {
   hasLoaded?: boolean
 }
 
+const track: Track<Props, State> = _track
+
+@screenTrack({
+  context_screen: Schema.PageNames.ConsignmentsOverView,
+  context_screen_owner_type: Schema.OwnerEntityTypes.Consignment,
+})
 export default class Info extends React.Component<Props, State> {
   constructor(props) {
     super(props)
@@ -118,16 +120,41 @@ export default class Info extends React.Component<Props, State> {
       updateSubmission(this.state, this.state.submission_id)
     } else if (this.state.artist) {
       const submission = await createSubmission(this.state)
-      this.setState({ submission_id: submission.id })
+      this.setState({ submission_id: submission.id }, () => {
+        this.submissionDraftCreated()
+      })
     }
+  }
+
+  @track((props, state) => ({
+    action_type: Schema.ActionTypes.Success,
+    action_name: Schema.ActionNames.ConsignmentDraftCreated,
+    owner_id: state.submission_id,
+    owner_type: Schema.OwnerEntityTypes.Consignment,
+    owner_slug: state.submission_id,
+  }))
+  submissionDraftCreated() {
+    return null
   }
 
   submitFinalSubmission = async (setup: ConsignmentSetup) => {
     this.setState(setup, async () => {
       await this.updateLocalStateAndMetaphysics()
       await AsyncStorage.removeItem(consignmentsStateKey)
+      this.submissionDraftSubmitted()
       this.exitModal()
     })
+  }
+
+  @track((props, state) => ({
+    action_type: Schema.ActionTypes.Success,
+    action_name: Schema.ActionNames.ConsignmentSubmitted,
+    owner_id: state.submission_id,
+    owner_type: Schema.OwnerEntityTypes.Consignment,
+    owner_slug: state.submission_id,
+  }))
+  submissionDraftSubmitted() {
+    return null
   }
 
   exitModal = () => SwitchBoard.dismissModalViewController(this)