+++ src/lib/Components/Consignments/Screens/Welcome.tsx
@@ -13,6 +13,7 @@ import {
 import { Router } from "lib/utils/router"
 import SwitchBoard from "../../../NativeModules/SwitchBoard"
 
+import { Schema, screenTrack, Track } from "lib/utils/track"
 import Circle from "../Components/CircleImage"
 import ConsignmentBG from "../Components/ConsignmentBG"
 import { Button } from "../Components/FormElements"
@@ -24,6 +25,10 @@ interface Props extends ViewProperties {
   route: Route
 }
 
+@screenTrack({
+  context_screen: Schema.PageNames.ConsignmentsWelcome,
+  context_screen_owner_type: Schema.OwnerEntityTypes.Consignment,
+})
 export default class Welcome extends React.Component<Props, null> {
   goTapped = () => this.props.navigator.push({ component: Overview })