+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 -   Minor: change text colour for provenance placeholder in consignments - maxim
 -   Don’t rely on user of the Gene VC to specify filter defaults, which could lead to a crash - alloy
 -   QA for inbox zero on ipad - maxim
+-   Add analytics to consignments - orta
+-   Fixes for metaphysics changes to consignment submission - orta
+-   Implement selected state for images in consignments photo set - maxim
+-   Support sending images for Consignments in production - orta
 
 ### 1.4.0-beta.8